run forge fmt and minor refactor

+++ lib/bonsai-lib-sol/src/BonsaiCheats.sol
@@ -25,17 +25,16 @@ import "forge-std/Base.sol";
 
 /// @notice A base contract for testing a Bonsai callback receiver contract
 abstract contract BonsaiCheats is StdCheatsSafe, CommonBase {
-
     using Strings2 for bytes;
 
     /// @notice Returns the journal resulting from running the guest with @imageId using @input.
     function queryImageOutput(bytes32 imageId, bytes memory input) internal returns (bytes memory) {
         string[] memory imageRunnerInput = new string[](6);
-        uint i = 0;
-        imageRunnerInput[i++] = 'cargo';
-        imageRunnerInput[i++] = 'run';
-        imageRunnerInput[i++] = '-q';
-        imageRunnerInput[i++] = 'query';
+        uint256 i = 0;
+        imageRunnerInput[i++] = "cargo";
+        imageRunnerInput[i++] = "run";
+        imageRunnerInput[i++] = "-q";
+        imageRunnerInput[i++] = "query";
         imageRunnerInput[i++] = abi.encodePacked(imageId).toHexString();
         imageRunnerInput[i++] = input.toHexString();
         return vm.ffi(imageRunnerInput);
@@ -44,11 +43,11 @@ abstract contract BonsaiCheats is StdCheatsSafe, CommonBase {
     /// @notice Returns the image id of the guest with the specified name.
     function queryImageId(string memory binaryName) internal returns (bytes32) {
         string[] memory imageRunnerInput = new string[](5);
-        uint i = 0;
-        imageRunnerInput[i++] = 'cargo';
-        imageRunnerInput[i++] = 'run';
-        imageRunnerInput[i++] = '-q';
-        imageRunnerInput[i++] = 'query';
+        uint256 i = 0;
+        imageRunnerInput[i++] = "cargo";
+        imageRunnerInput[i++] = "run";
+        imageRunnerInput[i++] = "-q";
+        imageRunnerInput[i++] = "query";
         imageRunnerInput[i++] = binaryName;
         return bytes32(vm.ffi(imageRunnerInput));
     }
